improve readability according to review

+++ include/iota/types/trinary.hpp
@@ -38,10 +38,10 @@ using Trytes = std::string;
 
 /**
  * @return the index of the tryte character in the tryte alphabet
- * the range for valid tryte characters is [0, 26], for invalid
- * characters an index larger than 26 is returned
+ * the range for valid tryte characters is [0, 26], 
+ * for invalid characters -1 is returned
  */
-size_t tryteIndex(const char& tryte);
+int8_t tryteIndex(const char& tryte);
 
 bool isValidTryte(const char& tryte);
